[ML-5672] simplify docker/travis setup
Similar to databricks/spark-deep-learning@846bfc0 we use docker/docker-compose to describe the test env and hence simplify the content in .travis.yml.

A major difference between this PR and databricks/spark-deep-learning@846bfc0 is that we use conda to activate the env instead of updating env variables directly. This provides better compatibility. But we need to use interactive shell to load the profile (bash -i).

We also mount "~/.ivy2" to avoid unnecessary download in sbt build.